test: keep authority regression under quality limits
